+++ windi.config.ts
@@ -66,5 +66,6 @@ export default defineConfig({
       "hover:bg-gray-200 data-focused:bg-gray-100 " +
       "data-disabled:(pointer-events-none text-gray-400) " +
       "mark:(underline text-[unset] bg-[unset])",
+    "solid-select-list-placeholder": "px-4 py-2 cursor-default select-none",
   },
 });